+++ harness/determined/cli/cli.py
@@ -24,6 +24,7 @@ import determined.cli
 from determined.cli import render
 from determined.cli.agent import args_description as agent_args_description
 from determined.cli.checkpoint import args_description as checkpoint_args_description
+from determined.cli.dev import args_description as dev_args_description
 from determined.cli.experiment import args_description as experiment_args_description
 from determined.cli.job import args_description as job_args_description
 from determined.cli.master import args_description as master_args_description
@@ -165,6 +166,7 @@ all_args_description = (
     + workspace_args_description
     + auth_args_description
     + oauth_args_description
+    + dev_args_description
 )
